+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Introduce TitleRenderer for customizable column titles ( by @doonfrs )
 * Implement custom loading widget support and add Loading Options feature to documentation and demo ( by @doonfrs )
 * Introduce Percentage Column Type ( by @doonfrs )
+* add decimalInput option to percentage column type ( by @doonfrs )
 
 ## [1.6.5] - 2025. 03. 22